// src/app/pages/permisotaxis/components/permisotaxis-table/permisotaxis.interface.ts
export interface PermisotaxisInterface {
   idpermisotaxi?: number;
   numero?: string;
   status?: string;
   fechaAlta?: string;
   vigencia?: string;
   liquidez?: number;
   liquidezDom?: number;
   propietario?: number;
   baja?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}